Update Editor.php
处理旧文章编辑功能（后台部分）

+++ Editor/Editor.php
@@ -10,15 +10,24 @@ class Editor {
 	function config_loaded($config) {
 		if(isset($_GET['edit'])) {
 			/* show editor */
+			$name = trim($_GET['edit']);
+			$file = CONTENTS_DIR.'/'.$name.'.md';
+			$content = 'date:' . date('Y-m-d H:i:s').PHP_EOL.'title:';
+			$newfile = true;
+			if($name != '' && file_exists($file)){
+				$content = file_get_contents($file);
+				$newfile = false;
+			}
 			include_once PLUGINS_DIR."/Editor/Editor.html";
 			die();
 		}
 
-		if(isset($_POST['editorname']) && isset($_POST['editortext'])) {
-			$name = DIRECTORY_SEPARATOR == '\\' ? iconv('utf-8', 'gbk', $_POST['editorname']) : $_POST['editorname'];
+		if(isset($_POST['name']) && isset($_POST['editortext'])) {
+			$newfile = $_POST['newfile'] == 'true';
+			$name = DIRECTORY_SEPARATOR == '\\' ? iconv('utf-8', 'gbk', $_POST['name']) : $_POST['name'];
 			$name = preg_replace('/[\/\\\<\>\*\?]/m','_',$name);
 			$file = CONTENTS_DIR.'/'.$name.'.md';
-			if(file_exists($file)){
+			if($newfile && file_exists($file)){
 				echo "<meta http-equiv=\"content-type\" content=\"text/html; charset=utf-8\"><script>alert('已经有一篇名为 “".$name."” 的文章了，还请换个名字保存！');window.location.href='?edit';</script>";
 				die();
 			}else{
